fix(app/courses/[course]/section/[section]/lesson/[lesson]/components): eslint errors
